Add authorization and authentication

+++ server/routes/posts.js
@@ -2,6 +2,7 @@ const express = require("express");
 const multer = require("multer");
 
 const Post = require('../models/post.model');
+const checkAuth = require('../middleware/check-auth');
 
 const router = express.Router();
 
@@ -26,12 +27,13 @@ const storage = multer.diskStorage({
     }
 });
 
-router.post('', multer({storage: storage}).single('image'), (req, res) => {
+router.post('', checkAuth, multer({ storage: storage }).single('image'), (req, res) => {
     const url = req.protocol + '://' + req.get("host");
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -45,7 +47,7 @@ router.post('', multer({storage: storage}).single('image'), (req, res) => {
 
 });
 
-router.put('/:id', multer({storage: storage}).single('image'), (req, res) => {
+router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req, res) => {
     let imagePath = req.body.imagePath;
     if (req.file) {
         const url = req.protocol + '://' + req.get("host");
@@ -55,13 +57,20 @@ router.put('/:id', multer({storage: storage}).single('image'), (req, res) => {
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath
+        imagePath,
+        creator: req.userData.userId
     });
 
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-        res.status(200).json({
-            message: 'Update successful',
-        })
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if (result.modifiedCount > 0) {
+            res.status(200).json({
+                message: 'Update successful',
+            });
+        } else {
+            res.status(401).json({
+                message: 'Not authorized!',
+            });
+        }
     })
         .catch(error => {
             console.log(error);
@@ -102,10 +111,16 @@ router.get('/:id', (req, res) => {
     })
 });
 
-router.delete('/:id', (req, res) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({message: 'Post deleted'});
+router.delete('/:id', checkAuth, (req, res) => {
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.deletedCount > 0) {
+            res.status(200).json({ message: 'Post deleted' });
+        } else {
+            res.status(401).json({
+                message: 'Not authorized!',
+            });
+        }
+
     });
 });
